requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosmtpd==1.4.3
alabaster==0.7.12
alembic==1.8.1.dev0
anyio==3.6.2
appdirs==1.4.4
arandr==0.1.11
argon2-cffi==21.1.0
arrow==1.3.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
async-lru==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.1
atpublic==2.3
attrs==22.2.0
authheaders==0.15.2
authres==1.2.0
autopep8==2.0.1
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.2
beniget==0.4.1
binaryornot==0.4.3
black==23.1.0
bleach==5.0.1
blinker==1.5
Bottleneck==1.3.5
breezy==3.3.2
Brlapi==0.8.4
Brotli==1.0.9
bytecode==0.14.0
cbor2==5.4.6
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
coloredlogs==15.0.1
colormath==3.0.0
configobj==5.0.8
contourpy==1.0.7
cookiecutter==1.7.3
coverage==6.5.0
cryptography==38.0.4
css-parser==1.0.8
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.32
datasets==3.5.0
dbus-python==1.3.2
debugpy==1.6.3+git20221103.a2a3328
decorator==5.1.1
deepTools==3.5.1
deeptoolsintervals==0.1.9
defusedxml==0.7.1
Deprecated==1.2.13
devscripts==2.23.4+deb12u1
diff-match-patch==20200713
dill==0.3.8
distlib==0.3.6
distro==1.8.0
distro-info==1.5+deb12u1
Django==3.2.19
django-filter==23.1
django-redis==5.2.0
django-tables2==2.4.1
djangosaml2==1.5.5
dkimpy==1.1.4
dnspython==2.3.0
docstring-to-markdown==0.11
docutils==0.19
dulwich==0.21.2
elementpath==2.5.3
entrypoints==0.4
et-xmlfile==1.0.1
executing==1.2.0
fail2ban==1.0.2
falcon==3.1.1
fastbencode==0.2
fasteners==0.17.3
fastimport==0.9.14
fastjsonschema==2.16.3
filelock==3.9.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
flufl.bounce==4.0
flufl.i18n==3.0.1
flufl.lock==5.0.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.12.0
future==0.18.2
gajim==1.7.3
gast==0.5.2
GDAL==3.6.2
gevent==22.10.2
git-filter-repo==2.38.0
google-api_core==2.25.1
google-genai==1.24.0
gpg==1.18.0
greenlet==2.0.2
gssapi==1.8.2
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
h2==4.1.0
helpdev==0.7.1
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
hyperframe==6.0.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.1.2
inflection==0.3.1
iniconfig==1.1.1
intervaltree==3.0.2
iotop==0.6
ipykernel==6.17.0
ipython==8.5.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.1
isoduration==20.11.0
isort==5.6.4
itsdangerous==2.1.2
jaraco.classes==3.2.1
jdcal==1.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==0.8.9
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
json5==0.12.0
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.3
keyring==23.9.3
kiwisolver==0.0.0
launchpadlib==1.11.0
lazr.config==2.2.3
lazr.delegates==2.0.4
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
libsass==0.22.0
llvmlite==0.39.1
lockfile==0.12.2
logilab-common==1.9.8
louis==3.24.0
lxml==4.9.2
lz4==4.0.2+dfsg
lzstring==1.0.4
mailman==3.3.8
mailmanclient==3.3.5
Mako==1.2.4.dev0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mercurial==6.3.2
merge3==0.0.8
mergedeep==1.3.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
msoffcrypto-tool==5.0.0
multidict==6.3.2
multiprocess==0.70.16
multiqc==1.14
mypy-extensions==0.4.3
nbclient==0.7.2
nbconvert==6.5.3
nbformat==5.5.0
nbxmpp==4.2.2
nest-asyncio==1.5.4
networkx==2.8.8
nose==1.3.7
notebook==6.4.12
notebook_shim==0.2.4
ntpsec==1.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.24.2
numpydoc==1.5.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
overrides==7.7.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parsimonious==0.10.0
parso==0.8.3
passlib==1.7.4
Paste==3.5.2
pathspec==0.11.0
patiencediff==0.2.13
patsy==0.5.3
pbr==5.10.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.0
plotly==5.4.0
pluggy==1.0.0+repack
ply==3.11
precis-i18n==1.0.5
prometheus-client==0.16.0
pronouncing==0.2.0
prompt-toolkit==3.0.36
propcache==0.3.1
protobuf==4.21.12
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.0.0
py==1.11.0
py2bit==0.3.1
pyarrow==19.0.1
pyasn1==0.4.8
pyBigWig==0.3.18
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.2
pydevd==2.9.5
pydocstyle==6.2.3
pydot==1.4.2
pyflakes==2.5.0
PyGithub==1.55
Pygments==2.14.0
PyGObject==3.42.2
pygraphviz==1.7
pyinotify==0.9.6
PyJWT==2.6.0
pykerberos==1.1.14
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
pymol==2.5.0
PyNaCl==1.5.0
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-sip==13.4.1
PyQtWebEngine==5.15.6
pyrsistent==0.18.1
pysam==0.20.0
pysaml2==7.0.1
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pysvn==1.9.15
pytest==7.2.1
python-apt==2.6.0
python-axolotl==0.2.3
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.8.2
python-debian==0.1.49
python-debianbts==4.0.1
python-dotenv==0.21.0
python-json-logger==3.3.0
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-magic==0.4.26
python-memcached==1.58
python-mimeparse==1.6.0
python-rapidjson==1.4
python-redmine==2.3.0
python-slugify==4.0.0
python3-saml==1.12.0
pythran==0.11.0
pytoolconfig==1.2.5
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
QDarkStyle==3.1
qrcode==7.4.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.3.0
qtsass==0.3.2
redis==4.3.4
referencing==0.36.2
regex==2022.10.31
rencode==1.0.6
reportbug==12.0.0
repoze.who==2.2
requests==2.32.3
requests-toolbelt==0.10.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987==1.3.8
rich==13.3.1
rich-click==1.6.0
roman==3.3
rope==1.7.0
rpds-py==0.24.0
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safetensors==0.5.3
scikit-learn==1.2.1
scipy==1.10.1
SCons==4.4.0
scour==0.38.2
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.3
sentry-sdk==1.9.10
setproctitle==1.3.1
simplejson==3.18.3
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
spectra==0.0.11
Sphinx==5.3.0
spyder==5.4.2
spyder-kernels==2.4.2
SQLAlchemy==1.4.46
sqlparse==0.4.2
stack-data==0.6.2
sympy==1.13.1
systemd-python==235
tables==3.7.0
tablib==3.4.0
tabulate==0.8.9
Tempita==0.5.2
tenacity==8.2.1
terminado==0.17.0
textdistance==4.5.0
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss2==1.2.1
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.11.7
torch==2.6.0
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.0
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
ufoLib2==0.14.0
ujson==5.7.0
Unidecode==1.3.6
unidiff==0.7.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1+ds
wadllib==1.3.6
watchdog==2.2.1
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.8.0
Werkzeug==2.2.2
widgetsnbextension==2.0.0
wrapt==1.14.1
wurlitzer==3.0.3
xdg==5
xlrd==1.2.0
xlwt==1.3.0
xmlschema==1.10.0
xmlsec==1.3.13
xxhash==3.5.0
yapf==0.32.0
yarl==1.19.0
zipp==1.0.0
zope.component==5.1.0
zope.configuration==4.4.1
zope.event==4.4
zope.hookable==5.1.0
zope.i18nmessageid==5.0.1
zope.interface==5.5.2
zope.schema==7.0.1
